renamed files and fix useParam in profile information
