+++ src/pages/Landing.test.tsx
@@ -2,20 +2,19 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
 import { store } from '../app/store';
-import Feed from './feed';
+import Landing from './Landing';
 import { BrowserRouter } from 'react-router-dom';
 
-describe('Feed testing', () => {
+describe('The landing page', () => {
 
-  it('should show Create Post', ()=>{
+  it('should show Reverb', ()=>{
     const { getByText } = render(
       <Provider store={store}>
           <BrowserRouter>
-            <Feed/>
+            <Landing/>
           </BrowserRouter>
       </Provider>
     );
-    expect(getByText("Create Post")).toBeInTheDocument();
+    expect(getByText("Reverb")).toBeInTheDocument();
   })
-
 });
